requirements.txt
discord.py[voice]>=2.3.2
yt-dlp>=2023.12.30
PyNaCl>=1.5.0 
ollama>=0.1.6
openai>=1.0.0